+++ src/ScottPlot4/ScottPlot.Tests/PlotTypes/SignalXY.cs
@@ -41,7 +41,34 @@ namespace ScottPlotTests.PlotTypes
             }
 
             var plt = new ScottPlot.Plot(500, 350);
-            Assert.Throws<ArgumentException>(() => { plt.AddSignalXY(xs, ys); });
+            plt.AddSignalXY(xs, ys);
+            Assert.Throws<InvalidOperationException>(() => { plt.Render(); });
+        }
+
+        [Test]
+        public void Test_OutsideOfRenderLimitsNotAllXsAscend_ValidateDataDoesNotThrowException()
+        {
+            int pointCount = 1000;
+            int minRenderIndex = 100;
+            int maxRenderIndex = 200;
+            double[] ys = new double[pointCount];
+            double[] xs = new double[pointCount];
+            for (int i = 0; i < minRenderIndex; i++)
+            {
+                ys[i] = double.MaxValue;
+                xs[i] = double.MaxValue;
+            }
+            for (int i = minRenderIndex; i <= maxRenderIndex; i++)
+            {
+                ys[i] = i;
+                xs[i] = i;
+            }
+
+            var plt = new ScottPlot.Plot(500, 350);
+            var signalXY = plt.AddSignalXY(xs, ys);
+            signalXY.MinRenderIndex = minRenderIndex;
+            signalXY.MaxRenderIndex = maxRenderIndex;
+            Assert.DoesNotThrow(() => { plt.Render(); });
         }
 
         [Test]